+++ tests/ServiceClientFactory.test.js
@@ -1,9 +1,19 @@
 import nock from 'nock';
 import mockFs from 'mock-fs';
+import { NotFoundError } from 'json-api-error';
 import ServiceClientFactory from '../src/ServiceClientFactory';
 
 describe('ServiceClientFactory', () => {
-  beforeAll(() => {
+  jest.mock('../src/logger', () => ({
+    error: jest.fn(),
+  }));
+
+  afterEach(() => {
+    mockFs.restore();
+    nock.restore();
+  });
+
+  it('can execute request to targeted service correctly', async () => {
     mockFs({
       'service-env.json': JSON.stringify({
         id: 'order-service',
@@ -17,17 +27,7 @@ describe('ServiceClientFactory', () => {
         }],
       }),
     });
-  });
 
-  afterEach(() => {
-    mockFs.restore();
-  });
-
-  afterEach(() => {
-    nock.restore();
-  });
-
-  it('can execute request to targeted service correctly', async () => {
     nock('http://storage-service:3000')
       .get('/healthz')
       .reply(200, {
@@ -42,5 +42,36 @@ describe('ServiceClientFactory', () => {
     });
 
     expect(response.status).toBe(200);
+  });
+
+  it('throw NotFoundError if the target was not found', async () => {
+    mockFs({
+      'service-env.json': JSON.stringify({
+        id: 'order-service',
+        name: 'order service',
+        dependencies: [{
+          id: 'storage-service',
+          name: 'storage service',
+          category: 'storage',
+          endpoint: 'http://storage-service:3000',
+          version: '1.0.0',
+        }],
+      }),
+    });
+
+    const orderServiceClient = await ServiceClientFactory.create('order-service');
+
+    let error;
+
+    try {
+      await orderServiceClient.request({
+        method: 'get',
+        url: '/healthz',
+      });
+    } catch (err) {
+      error = err;
+    }
+
+    expect(error).toBeInstanceOf(NotFoundError);
   });
 });